refactor : Result DTO 위치 변경
